pyproject.toml
[project]
name = "brief-agent"
version = "0.1.0"
description = "企业级多模态知识内生 Agent 平台 - AI 简报制片人 (Insights-to-Video Pipeline)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.14.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "sqlite-utils>=3.37",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.1",
    "apscheduler>=3.10.4",
    "loguru>=0.7.2",
    "pydantic>=2.9.0",
    "pytest>=9.0.2",
    "langchain-openapi>=0.1.1",
    "langchain-openai>=1.1.7",
    "sqlmodel>=0.0.31",
    "dashscope>=1.25.9",
    "python-pptx>=1.0.2",
]

[project.scripts]
brief = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
